Citation popover: status dot per row + quick report-flag
Two changes that make citations more legible/actionable inline:

1. Status indicator (green/yellow/red) on each source in the popover,
   matching /sources. Reader sees the citation's verification state
   without leaving the page.

2. Report-flag (⚑) at the right edge of each row. Click opens the
   source-report issue template in a new tab with the source URL and
   title pre-filled — frictionless escalation when something looks off.

To keep the popover and /sources in sync, the status logic + StatusDot
component lift out of components/Sources.tsx into a new shared module
src/lib/sourceStatus.tsx. Both views now classify Broken / Overdue /
Verified the same way and reuse the same dot styling. ReportFlag also
lives there since it's the other place we point at the source-report
template.

The popover row gains a small layout change: status dot on the left,
the existing label/tier/date link in the middle, the flag on the right
— a 3-column flex row inside each item. Nested-anchor invalid HTML
fixed in the process; the dot and flag are independent click targets.
